change p_eight_sad16x16 and p_eight_sad8x8 from 32bit to 16bit

+++ Source/Lib/Common/ASM_AVX2/EbComputeSAD_AVX2.h
@@ -139,12 +139,12 @@ extern "C" {
         uint32_t  *p_best_sad16x16,
         uint32_t  *p_best_mv8x8,
         uint32_t  *p_best_mv16x16,
-        uint32_t   p_eight_sad16x16[16][8],
-        uint32_t   p_eight_sad8x8[64][8]);
+        uint16_t   p_eight_sad16x16[16][8],
+        uint16_t   p_eight_sad8x8[64][8]);
 
     void ext_eigth_sad_calculation_nsq_avx2(
-        uint32_t   p_sad8x8[64][8],
-        uint32_t   p_sad16x16[16][8],
+        uint16_t   p_sad8x8[64][8],
+        uint16_t   p_sad16x16[16][8],
         uint32_t   p_sad32x32[4][8],
         uint32_t  *p_best_sad64x32,
         uint32_t  *p_best_mv64x32,
@@ -169,7 +169,7 @@ extern "C" {
         uint32_t   mv);
 
     void ext_eight_sad_calculation_32x32_64x64_avx2(
-        uint32_t  p_sad16x16[16][8],
+        uint16_t  p_sad16x16[16][8],
         uint32_t *p_best_sad32x32,
         uint32_t *p_best_sad64x64,
         uint32_t *p_best_mv32x32,
