+++ Source/Lib/Common/Codec/EbMotionEstimation.h
@@ -386,16 +386,16 @@ void interpolate_search_region_AVC_chroma(
         uint32_t *p_best_sad16x16,
         uint32_t *p_best_mv8x8,
         uint32_t *p_best_mv16x16,
-        uint32_t p_eight_sad16x16[16][8],
-        uint32_t p_eight_sad8x8[64][8]);
+        uint16_t p_eight_sad16x16[16][8],
+        uint16_t p_eight_sad8x8[64][8]);
 
     /****************************************************
     Calcualte SAD for Rect H, V and H4, V4 partitions
     and update its Motion info if the result SAD is better
     ****************************************************/
     extern void ext_eigth_sad_calculation_nsq_c(
-        uint32_t p_sad8x8[64][8],
-        uint32_t p_sad16x16[16][8],
+        uint16_t p_sad8x8[64][8],
+        uint16_t p_sad16x16[16][8],
         uint32_t p_sad32x32[4][8],
         uint32_t *p_best_sad64x32,
         uint32_t *p_best_mv64x32,
@@ -425,7 +425,7 @@ void interpolate_search_region_AVC_chroma(
     the best SAD+MV
     *******************************************/
     extern void ext_eight_sad_calculation_32x32_64x64_c(
-        uint32_t p_sad16x16[16][8],
+        uint16_t p_sad16x16[16][8],
         uint32_t *p_best_sad32x32,
         uint32_t *p_best_sad64x64,
         uint32_t *p_best_mv32x32,
