+++ Source/Lib/Common/Codec/EbMotionEstimationContext.h
@@ -430,8 +430,8 @@ extern "C" {
         uint8_t                     *p_best_nsq64x64;
         uint16_t                     *p_eight_pos_sad16x16;
         EB_ALIGN(64) uint32_t         p_eight_sad32x32[4][8];
-        EB_ALIGN(64) uint32_t         p_eight_sad16x16[16][8];
-        EB_ALIGN(64) uint32_t         p_eight_sad8x8[64][8];
+        EB_ALIGN(64) uint16_t         p_eight_sad16x16[16][8];
+        EB_ALIGN(64) uint16_t         p_eight_sad8x8[64][8];
         EbBitFraction               *mvd_bits_array;
         uint64_t                      lambda;
         uint8_t                       hme_search_type;
